// N/client/src/module/Admin/admin.validator.tsx
export const BannerType = {
    Homepage: 'homepage',
    Category: 'category'
} as const

export type BannerType = typeof BannerType[keyof typeof BannerType]

export interface createBannerProps {
    title: string,
    type: BannerType,
    image?: string | null,
    startAt: dayjs.Dayjs,
    endAt: dayjs.Dayjs,
    isActive?: boolean | null,
    priority?: number | null
}

export const BannerValidationDTO = yup.object().shape({
    title: yup.string().required("Title is required."),
    startAt: yup.mixed<dayjs.Dayjs>().required('Start date is required'),
    endAt: yup.mixed<dayjs.Dayjs>().required('End date is required'),
    type: yup.mixed<BannerType>().oneOf(Object.values(BannerType), "Invalid Type").required("Type is required"),
    image: yup.string().nullable(),
    isActive: yup.boolean().nullable(),
    priority: yup.number().nullable(),
})